--- a/src/caches/extrapolation_caches.jl
+++ b/src/caches/extrapolation_caches.jl
@@ -229,12 +229,16 @@ function alg_cache(alg::ImplicitEulerExtrapolation,u,rate_prototype,::Type{uElty
     linsolve_tmps[i] = zero(rate_prototype)
   end
 
-  linsolve_el = alg.linsolve(Val{:init},uf,u)
-  linsolve = Array{typeof(linsolve_el),1}(undef, Threads.nthreads())
-  linsolve[1] = linsolve_el
+  linprob = LinearProblem(W,vec(linsolve_tmps[1]); u0=vec(k_tmps[1]))
+  linsolve1 = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
+
+  linsolve = Array{typeof(linsolve1),1}(undef, Threads.nthreads())
+  linsolve[1] = linsolve1
   for i=2:Threads.nthreads()
-    linsolve[i] = alg.linsolve(Val{:init},uf,u)
+    linprob = LinearProblem(W,vec(linsolve_tmps[i]); u0=vec(k_tmps[i]))
+    linsolve[i] = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
   end
+
   res = uEltypeNoUnits.(zero(u))
   grad_config = build_grad_config(alg,f,tf,du1,t)
   jac_config = build_jac_config(alg,f,uf,du1,uprev,u,du1,du2)
@@ -728,11 +732,14 @@ function alg_cache(alg::ImplicitDeuflhardExtrapolation,u,rate_prototype,::Type{u
     linsolve_tmps[i] = zero(rate_prototype)
   end
 
-  linsolve_el = alg.linsolve(Val{:init},uf,u)
-  linsolve = Array{typeof(linsolve_el),1}(undef, Threads.nthreads())
-  linsolve[1] = linsolve_el
+  linprob = LinearProblem(W,vec(linsolve_tmps[1]); u0=vec(k_tmps[1]))
+  linsolve1 = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
+
+  linsolve = Array{typeof(linsolve1),1}(undef, Threads.nthreads())
+  linsolve[1] = linsolve1
   for i=2:Threads.nthreads()
-    linsolve[i] = alg.linsolve(Val{:init},uf,u)
+    linprob = LinearProblem(W,vec(linsolve_tmps[i]); u0=vec(k_tmps[i]))
+    linsolve[i] = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
   end
   grad_config = build_grad_config(alg,f,tf,du1,t)
   jac_config = build_jac_config(alg,f,uf,du1,uprev,u,du1,du2)
@@ -1016,11 +1023,14 @@ function alg_cache(alg::ImplicitHairerWannerExtrapolation,u,rate_prototype,::Typ
     linsolve_tmps[i] = zero(rate_prototype)
   end
 
-  linsolve_el = alg.linsolve(Val{:init},uf,u)
-  linsolve = Array{typeof(linsolve_el),1}(undef, Threads.nthreads())
-  linsolve[1] = linsolve_el
+  linprob = LinearProblem(W,vec(linsolve_tmps[1]); u0=vec(k_tmps[1]))
+  linsolve1 = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
+
+  linsolve = Array{typeof(linsolve1),1}(undef, Threads.nthreads())
+  linsolve[1] = linsolve1
   for i=2:Threads.nthreads()
-    linsolve[i] = alg.linsolve(Val{:init},uf,u)
+    linprob = LinearProblem(W,vec(linsolve_tmps[i]); u0=vec(k_tmps[i]))
+    linsolve[i] = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
   end
   grad_config = build_grad_config(alg,f,tf,du1,t)
   jac_config = build_jac_config(alg,f,uf,du1,uprev,u,du1,du2)
@@ -1186,11 +1196,14 @@ function alg_cache(alg::ImplicitEulerBarycentricExtrapolation,u,rate_prototype,:
     linsolve_tmps[i] = zero(rate_prototype)
   end
 
-  linsolve_el = alg.linsolve(Val{:init},uf,u)
-  linsolve = Array{typeof(linsolve_el),1}(undef, Threads.nthreads())
-  linsolve[1] = linsolve_el
+  linprob = LinearProblem(W,vec(linsolve_tmps[1]); u0=vec(k_tmps[1]))
+  linsolve1 = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
+
+  linsolve = Array{typeof(linsolve1),1}(undef, Threads.nthreads())
+  linsolve[1] = linsolve1
   for i=2:Threads.nthreads()
-    linsolve[i] = alg.linsolve(Val{:init},uf,u)
+    linprob = LinearProblem(W,vec(linsolve_tmps[i]); u0=vec(k_tmps[i]))
+    linsolve[i] = init(linprob,alg.linsolve,alias_A=true,alias_b=true)
   end
   grad_config = build_grad_config(alg,f,tf,du1,t)
   jac_config = build_jac_config(alg,f,uf,du1,uprev,u,du1,du2)