@ChrisRackauckas 2021-12-11
Test case is:

```julia
using OrdinaryDiffEq, StaticArrays, BenchmarkTools
function rober(du,u,p,t)
    y₁,y₂,y₃ = u
    k₁,k₂,k₃ = p
    @inbounds begin
    du[1] = -k₁*y₁+k₃*y₂*y₃
    du[2] =  k₁*y₁-k₂*y₂^2-k₃*y₂*y₃
    du[3] =  k₂*y₂^2
    end
    nothing
end
prob = ODEProblem(rober,[1.0,0.0,0.0],(0.0,1e5),[0.04,3e7,1e4])
@btime solve(prob,Rosenbrock23()) # 761.578 μs (633 allocations: 53.05 KiB)
```

Only the Rosenbrock and Extrapolation methods will be tedious, the rest should already be handled by the Newton part.