@ChrisRackauckas 2021-12-20
Tests are passing, so now I'm just going to refactor a bit to condense the code before merging 🎉