@ChrisRackauckas 2021-12-20
Benchmarks look to be comfortably a wash even though it allocates more.

```julia
using OrdinaryDiffEq, SnoopCompile, ForwardDiff

lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
end

u0 = [1.0;0.0;0.0]; tspan = (0.0,100.0);
prob = ODEProblem(lorenz,u0,tspan); alg = Rodas5();
tinf = @snoopi_deep ForwardDiff.gradient(u0 -> sum(solve(ODEProblem(lorenz,u0,tspan),alg)), u0)
tinf = @snoopi_deep ForwardDiff.gradient(u0 -> sum(solve(ODEProblem(lorenz,u0,tspan),alg)), u0)

# Before
InferenceTimingNode: 1.219960/3.583559 on Core.Compiler.Timings.ROOT() with 34 direct children
InferenceTimingNode: 1.064226/1.756057 on Core.Compiler.Timings.ROOT() with 13 direct children

# After
InferenceTimingNode: 1.257616/3.774263 on Core.Compiler.Timings.ROOT() with 33 direct children
InferenceTimingNode: 1.006660/1.756162 on Core.Compiler.Timings.ROOT() with 13 direct children

using OrdinaryDiffEq, SnoopCompile, ForwardDiff, BenchmarkTools

lorenz = (du,u,p,t) -> begin
        du[1] = 10.0(u[2]-u[1])
        du[2] = u[1]*(28.0-u[3]) - u[2]
        du[3] = u[1]*u[2] - (8/3)*u[3]
end

u0 = [1.0;0.0;0.0]; tspan = (0.0,100.0);
prob = ODEProblem(lorenz,u0,tspan); alg = Rodas5();
tinf = @snoopi_deep solve(ODEProblem(lorenz,u0,tspan),alg)
tinf = @snoopi_deep solve(ODEProblem(lorenz,u0,tspan),alg)
@btime solve(ODEProblem(lorenz,u0,tspan),alg)

@btime solve(ODEProblem(lorenz,u0,(0.0,10.0)),alg)

# Before
InferenceTimingNode: 1.060946/2.763417 on Core.Compiler.Timings.ROOT() with 8 direct children
InferenceTimingNode: 0.002333/0.002333 on Core.Compiler.Timings.ROOT() with 0 direct children
2.055 ms (14069 allocations: 781.14 KiB)
175.000 μs (1341 allocations: 75.97 KiB)

# After
InferenceTimingNode: 1.129433/3.072010 on Core.Compiler.Timings.ROOT() with 7 direct children
InferenceTimingNode: 0.002182/0.002182 on Core.Compiler.Timings.ROOT() with 0 direct children
2.221 ms (22155 allocations: 2.24 MiB)
162.900 μs (1917 allocations: 184.00 KiB)
```

ROBER Before after

<img width="298" alt="rober_before" src="https://user-images.githubusercontent.com/1814174/146838692-9640a202-7f97-432d-b275-8ad42020abcd.png">
<img width="299" alt="rober_after" src="https://user-images.githubusercontent.com/1814174/146838711-a7004bc9-5707-4b85-962b-ae3205721281.png">

Pollu Before after

<img width="299" alt="pollu_before" src="https://user-images.githubusercontent.com/1814174/146838740-ffa9c254-b0b4-438b-9b87-31707dfd74ab.png">

<img width="298" alt="pollu_after" src="https://user-images.githubusercontent.com/1814174/146838745-e60c84d4-25c0-4729-b88e-6c7149f4d15b.png">