@ChrisRackauckas 2021-12-20
https://github.com/SciML/DelayDiffEq.jl/pull/222
https://github.com/SciML/StochasticDiffEq.jl/pull/450

Downstream PRs are already ready.